csrf protection to journalsindex delete related #609

+++ src/Ojs/JournalBundle/Controller/JournalsIndexController.php
@@ -13,6 +13,7 @@ use Ojs\JournalBundle\Entity\JournalsIndex;
 use Ojs\JournalBundle\Form\JournalsIndexType;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
+use Symfony\Component\Security\Core\Exception\TokenNotFoundException;
 
 /**
  * JournalsIndex controller.
@@ -44,6 +45,8 @@ class JournalsIndexController extends Controller
         $grid = $this->get('grid')->setSource($source);
 
         $actionColumn = new ActionsColumn("actions", 'actions');
+        ActionHelper::setup($this->get('security.csrf.token_manager'));
+
         $rowAction[] = ActionHelper::showAction('manager_journals_indexes_show', 'id');
         $rowAction[] = ActionHelper::editAction('manager_journals_indexes_edit', 'id');
         $rowAction[] = ActionHelper::deleteAction('manager_journals_indexes_delete', 'id');
@@ -229,16 +232,23 @@ class JournalsIndexController extends Controller
     }
 
     /**
-     * Deletes a JournalsIndex entity.
-     *
+     * @param Request $request
+     * @param $id
+     * @return RedirectResponse
+     * @throws TokenNotFoundException
      */
-    public function deleteAction($id)
+    public function deleteAction(Request $request, $id)
     {
         $em = $this->getDoctrine()->getManager();
         $entity = $em->getRepository('OjsJournalBundle:JournalsIndex')->find($id);
         if (!$entity) {
             throw $this->createNotFoundException('notFound');
         }
+
+        $csrf = $this->get('security.csrf.token_manager');
+        $token = $csrf->getToken('manager_journals_indexes'.$id);
+        if($token!=$request->get('_token'))
+            throw new TokenNotFoundException("Token Not Found!");
         $em->remove($entity);
         $em->flush();
         $this->successFlashBag('successful.remove');
